+++ amplify/data/resource.ts
@@ -3,11 +3,13 @@ import { type ClientSchema, a, defineData } from '@aws-amplify/backend'
 const schema = a.schema({
   Todo: a
     .model({
+      id: a.id().required(),
       content: a.string(),
       executionDate: a.date(),
-      isDone: a.boolean(),
+      isDone: a.boolean().default(false),
+      prevId: a.integer(),
     })
-    .authorization((allow) => [allow.owner()]),
+    .authorization((allow) => [allow.owner()])
 })
 
 export type Schema = ClientSchema<typeof schema>
