+++ src/components/TodoList/useTodo.ts
@@ -1,9 +1,19 @@
 import type { Schema } from '@/amplify/data/resource'
 import { generateClient } from 'aws-amplify/api'
+import { useEffect, useState } from 'react'
 
 const client = generateClient<Schema>()
 
 export const useTodo = () => {
+  const [todos, setTodos] = useState<Array<Schema['Todo']['type']>>([])
+
+  useEffect(() => {
+    const sub = client.models.Todo.observeQuery().subscribe({
+      next: ({ items }) => setTodos([...items]),
+    })
+    return () => sub.unsubscribe()
+  }, [])
+
   const update = async (
     models: Pick<
       Schema['Todo']['type'],
@@ -30,5 +40,5 @@ export const useTodo = () => {
       console.error(errors)
     }
   }
-  return { update, remove }
+  return { todos, update, remove }
 }
